Fixed ticket 107 (https://sourceforge.net/p/paradoxgameconverters/tickets/107/).
Basically, the converter tool uses the save game file name as the name for the output mod, which the frontend later copies to the correct place.
But a couple of characters (whitespace, dash) are changed into underscores, which the frontend previously didn't account for, and thus didn't
find the output mod. This is now fixed.

Some work has also been done to make the codebase more testable and IOC container friendly, as I'm probably heading in that direction.