Updated pop conversion to 1.13
value "baseProd" needs to be added as available value for this converter since BaseProduction is now in game. 
BaseManpower is shown as value "manpower" and BaseTax as value "baseTax"
It needs to be checked if new trade goods dyes, tropical wood and silk are used correct by converter

U guys dont know me, but the calculation of income and manpower aswell as adding values for all those buildings for previous version was done by me for blayne for a grandcampaign, so u can trust me. I cant add baseProd as value myself because i am not a programmer.