As I told ya
We'll keep the decision, letting the player decide. This will be probably until we find a better compromise to any althist, like adding post-colonial dynamic cultures (e.g. Neo-Swedish) and making it so they get cores this way, but it'd flow up, or we have better ideas.

I edited it so that if the player doesn't check it by 1862, the decision won't be enacted anymore, as it'll make no sense to keep it there.