Reduce log warning spam
Country history contains some top-level tags that are not dates, reflecting intial-status conditions and other non-time-sensitive things. The parser, while looking for active leaders, was trying to interpret these as dates anyway, and throwing a good deal of warnings in the process.

This fix does not change functionality - the misparsed tags were immediately getting discarded anyway due to being interpreted (through error) as date 0.0.0, which is long before the cutoff used in this function.

Fixes #311.